feat: centralize framework style configuration for improved maintainability
