+++ src/components/home/device-section.component.ts
@@ -1,4 +1,5 @@
 import {BaseElement, Component, HTML} from "@ayu-sh-kr/dota-core";
+import {FrameworkStyleConfig} from "@dota/components/home/utils/device-preview.component.ts";
 
 
 @Component({
@@ -27,11 +28,4 @@ export class DeviceSectionComponent extends BaseElement {
       </section>
     `;
   }
-}
-
-export const FrameworkStyleConfig = {
-  react: "framework-logo hover:drop-shadow-[0px_10px_15px_rgba(34,211,238,0.8)]",
-  angular: "framework-logo hover:drop-shadow-[0px_10px_15px_rgba(220,38,38,0.8)]",
-  vue: "framework-logo hover:drop-shadow-[0px_10px_15px_rgba(22,163,74,0.8)]",
-  solid: "framework-logo hover:drop-shadow-[0px_10px_15px_rgba(147,51,234,0.8)]",
 }
